requirements.txt
dash==3.0.3
dash_bootstrap_components==2.0.2
dash_leaflet==1.1.1
geopandas==1.0.1
numpy==2.2.5
pandas==2.2.3
plotly==6.0.1
python-dotenv==1.1.0
Requests==2.32.3
rpy2==3.5.17
scikit_learn==1.6.1
scipy==1.15.2
Shapely==2.1.0
reportlab==4.4.1
selenium==4.32.0
webdriver-manager==4.0.2
folium==0.19.6
tabulate==0.9.0
types-requests==2.32.0.20250515
pylint==3.3.7
mypy==1.15.0
flake8==7.2.0
pydocstyle==6.3.0
matplotlib==3.10.3
